// crud-localstorage/src/models/employee.ts
export interface Employee {
  empId: number;
  name: string;
  dob: string;
  gender: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  pinCode: string;
  department: string;
}


// export class Employee {
//     constructor(
//         public empId: number,
//         public name: string,
//         public dob: string,
//         public gender: string,
//         public email: string,
//         public phone: string,
//         public address: string,
//         public city: string,
//         public state: string,
//         public pinCode: string,
//         public department: string
//     ) { }
// }